Fix OBS capture for versions 28+
This change addresses the issue of capturing OBS version 28+, and
addresses a series of issues found while trying to make a build.

The list of changes is as follows:
  - [capture] Change PrintWindow parameters to include PW_RENDERFULLCONTENT
      which is required to capture OBS. Tested OBS 27+ on Win11 and amarec.
  - [updater] Fix exception on failure to load local version.
  - [tensorflow] Adapt to latest API for model loader/call.
  - [PyQt] Adapt to latest API (fix missing int conversion).
  - [config] Rename config.ini to config.json since it is a JSON file.
  - [config] Rename defaults.ini to defaults.json since it is a JSON file.
  - [meta] Introduce a requirements.txt file just listing the dependencies.
  - [meta] Add .spec file for pyinstaller to generate build with required data.

Build command:  pyinstaller .\AutoSplit64.spec

+++ as64gui/widgets/split_list.py
@@ -14,12 +14,12 @@ class SplitListWidget(QtWidgets.QWidget):
     def paintEvent(self, event):
         painter = QtGui.QPainter()
 
-        split_height = event.rect().height() / self._display_count
+        split_height = int(event.rect().height() / self._display_count)
         split_width = event.rect().width()
         text_width = split_width - split_height
         x_pos = split_height
-        pixmap_dimension = split_height * 0.8
-        pixmap_spacing = (split_height - pixmap_dimension) / 2
+        pixmap_dimension = int(split_height * 0.8)
+        pixmap_spacing = int((split_height - pixmap_dimension) / 2)
 
         painter.begin(self)
